delete handler 기능 개선

+++ forum/pages/api/post/delete.js
@@ -1,11 +1,20 @@
 import { connectDB } from '@/util/database';
 import { ObjectId } from 'mongodb';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 export default async function handler(req, res) {
   if (req.method === 'POST') {
-    // console.log(req.body);
+    let session = await getServerSession(req, res, authOptions);
+
     const db = (await connectDB).db('forum');
-    let result = await db.collection('post').deleteOne({ _id: new ObjectId(req.body) });
-    res.status(200).json('삭제완료');
+    let myList = await db.collection('post').findOne({ _id: new ObjectId(req.body) });
+
+    if (myList.author === session.user.email) {
+      let result = await db.collection('post').deleteOne({ _id: new ObjectId(req.body) });
+      res.status(200).json('삭제완료');
+    } else {
+      return res.status(500).json('현재 유저가 작성한 글이 아님');
+    }
   }
 }
